clonedriveuploader/youtube_automation/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
assemblyai==0.42.0
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
decorator==5.2.1
diffusers==0.34.0
distro==1.9.0
docstring_parser==0.16
fastapi==0.104.1
filelock==3.18.0
fsspec==2025.5.1
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-aiplatform==1.104.0
google-cloud-bigquery==3.35.0
google-cloud-core==2.4.3
google-cloud-language==2.17.2
google-cloud-resource-manager==1.14.2
google-cloud-speech==2.33.0
google-cloud-storage==2.19.0
google-cloud-texttospeech==2.27.0
google-cloud-translate==3.21.1
google-cloud-videointelligence==2.16.2
google-cloud-vision==3.10.2
google-crc32c==1.7.1
google-genai==1.4.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.73.1
grpcio-status==1.73.1
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.4
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==2.1.5
moviepy==2.2.1
mpmath==1.3.0
networkx==3.5
numpy==2.3.1
oauthlib==3.3.1
openai==1.97.0
packaging==25.0
pillow==11.3.0
proglog==0.1.12
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
safetensors==0.5.3
setuptools==78.1.0
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
sympy==1.14.0
torch==2.9.0.dev20250720+cpu
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==14.2
zipp==3.23.0
